Module linear_sim::collision

Collision detection and resolution

Re-exports

pub use self::contact::Contact;

pub use self::proximity::Proximity;

pub use self::intersection::Intersection;

Modules

contact

intersection

proximity

Structs

Collision
Collision subsystem

Constants

CONTACT_DISTANCE
Distance criterea for contacts and collisions (TOI contacts).

OBJECT_KEY_MAX
The value of the dynamic bit also gives the maximum key value allowed for objects in the collision system.

RESTING_VELOCITY
Velocity for determination of persistent contacts.

Functions

lerp_object
Linear-interpolate an object with the given pseudovelocity

report_sizes