Module linear_sim::object

Simulation object entities

Structs

Dynamic
A free, bounded, massive object

Identifier
Each value identifies a unique object by the kind of object and its key

Key
A VecMap index

Nodetect
A free, massive object

Static
A fixed, bounded object

Enums

Kind
Object kind

Variant
One of object variants

Traits

Bounded
A bounded object

Inertial
Inertial object

Massive
A bounded object with mass

Object
Base object trait

Solid
A bounded object with material properties

Temporal
Object with time derivatives

Functions

report_sizes
Print object size information

Type Definitions

KeyType
Object keys are used to uniquely identify an object of a specific kind: static, dynamic, or nocollide.